get field request working like before
