+++ public/jira-oidc-helpers/makeFieldsRequest.ts
@@ -1,31 +1,32 @@
 import { hasValidAccessToken } from "./auth";
-import { Config } from "./types";
+import { Config, FieldsRequest } from "./types";
 
 export function fetchJiraFields(config: Config) {
     return () => {
         return config.requestHelper(`/api/3/field`);
     };
 }
-
-export function makeFieldsRequest(config: Config) {
+export function makeFieldsRequest(config: Config, setFieldsRequest: (req: FieldsRequest) => void) {
     if (config.host === "jira" || hasValidAccessToken()) {
-    return fetchJiraFields(config)().then((fields) => {
-        const nameMap: Record<string, any> = {};
-        const idMap: Record<string, any> = {};
-        // @ts-ignore
-        fields.forEach((f) => {
+        const req = fetchJiraFields(config)().then((fields) => {
+            const nameMap: Record<string, any> = {};
+            const idMap: Record<string, any> = {};
             // @ts-ignore
-            idMap[f.id] = f.name;
-            // @ts-ignore
-            nameMap[f.name] = f.id;
-        });
-        console.log(nameMap);
+            fields.forEach((f) => {
+                // @ts-ignore
+                idMap[f.id] = f.name;
+                // @ts-ignore
+                nameMap[f.name] = f.id;
+            });
+            console.log(nameMap);
 
-        return {
-            list: fields,
-            nameMap: nameMap,
-            idMap: idMap,
-        };
-    });
-}
+            return {
+                list: fields,
+                nameMap: nameMap,
+                idMap: idMap,
+            };
+        });
+
+        setFieldsRequest(req)
+    }
 }
